requirements.txt
streamlit==1.40.2
opencv-python==4.6.0.66
numpy==1.26.4
Pillow==10.4.0